make it easier to create enum from "dumb" enums
see #93
